requirements.txt
google-generativeai
langchain
langchain-community
langchain-core
faiss-cpu
sentence-transformers
langgraph